add simple `targeted` function
This function follows the discussion laid out on
https://github.com/dask/dask/issues/2119.

This function simply changes the dask graph of the returned value, and does not
need to be used with a callback. However, this is something that would probably
be work adding since it could show exactly the dependencies which need to be run.