add another potential luigi interface
This commit provides a Targeted class which takes a function `f` and a list of
luigi targets, and returns a equally long list of dask delayed objects.

This commit also adds some useful functions for manipulating the dask tree.
These include

1. filter_tree
2. unfuse_match

In particular, unfuse_match will be useful for moving matching entries (e.g.
Targeted) function calls to the top level dictionary of the dask graph. This
will make it easy to process these matches.